create threading for qr_detection

+++ qrscan.py
@@ -1,7 +1,9 @@
-from camera import Camera
 import io
+from functools import partial
+from threading import Thread
 import numpy as np
 import cv2 as cv
+from camera import Camera
 
 camera = Camera()
 frame_size = (320, 240)
@@ -13,32 +15,39 @@ class StreamingOutput(object):
     def __init__(self):
         self.buffer = io.BytesIO()
         self.qrDetector = cv.QRCodeDetector()
+        self.bufferReady = True
 
     def write(self, buf):
-        if buf.startswith(b'\xff\xd8'):
-            # New frame
-            self.buffer.truncate()
-            # Reading QR
-            frame = self.buffer.getvalue()
-            npFrame = np.asarray(bytearray(frame))
-            if npFrame.any():
-                try:
-                    img = cv.imdecode(npFrame, 1)
-                    data, _, __ = self.qrDetector.detectAndDecode(img)
-                    if data:
-                        print (data)
-                    print ('ok')                      
-                except Exception as e:
-                    print (f'Qr detection error {e}')
-            print ('ok')
-            # Reset buffer pointer
-            self.buffer.seek(0)
+        if self.bufferReady:
+            if buf.startswith(b'\xff\xd8'):
+                # New frame
+                self.buffer.truncate()
+                # Reading QR
+                frame = self.buffer.getvalue()
+                t_read_qr = Thread(target = partial(self.read_qr, frame))
+                t_read_qr.start()
+                print ('ok')
+                # Reset buffer pointer
+                self.buffer.seek(0)
+            return self.buffer.write(buf)
 
-        return self.buffer.write(buf)
+    def read_qr(self, frame):
+        self.bufferReady = False
+        npFrame = np.asarray(bytearray(frame))
+        if npFrame.any():
+            try:
+                img = cv.imdecode(npFrame, 1)
+                data, _, __ = self.qrDetector.detectAndDecode(img)
+                if data:
+                    print (data)
+                print ('ok')                      
+            except Exception as e:
+                print (f'Qr detection error {e}')
+        self.bufferReady = True
 
 # Initialize streaming output object
 output = StreamingOutput()
 
 camera.start_camera(output, frame_size = frame_size, frame_rate = frame_rate)
-camera.wait_recording(20)
+camera.wait_recording(60)
 camera.stop_camera()
